Move commit module into show_commit
The commit module is only referenced by the show_commit module. This
change merges the commit module into the show_commit module and updates
the access modifiers to limit access to only the show_commit module.
