+++ src/show_commit/commit.rs
@@ -1,15 +1,10 @@
-mod file_stat;
-pub(crate) mod status;
-mod user;
-mod utils;
-
-use crate::commit::file_stat::FileStat;
-use crate::commit::user::User;
-use crate::commit::utils::load_commit_state;
+use crate::show_commit::file_stat::FileStat;
+use crate::show_commit::user::User;
+use crate::show_commit::util::load_commit_state;
 use chrono::{DateTime, Local};
 
 #[derive(Debug, PartialEq)]
-pub(crate) struct Commit {
+pub(super) struct Commit {
 	author: User,
 	body: Option<String>,
 	committer: User,
@@ -38,31 +33,31 @@ impl Commit {
 		}
 	}
 
-	pub(crate) fn from_commit_hash(hash: &str) -> Result<Self, String> {
+	pub(super) fn from_commit_hash(hash: &str) -> Result<Self, String> {
 		load_commit_state(hash).map_err(|e| String::from(e.message()))
 	}
 
-	pub(crate) fn get_author(&self) -> &User {
+	pub(super) fn get_author(&self) -> &User {
 		&self.author
 	}
 
-	pub(crate) fn get_committer(&self) -> &User {
+	pub(super) fn get_committer(&self) -> &User {
 		&self.committer
 	}
 
-	pub(crate) fn get_date(&self) -> &DateTime<Local> {
+	pub(super) fn get_date(&self) -> &DateTime<Local> {
 		&self.date
 	}
 
-	pub(crate) fn get_hash(&self) -> &String {
+	pub(super) fn get_hash(&self) -> &String {
 		&self.hash
 	}
 
-	pub(crate) fn get_body(&self) -> &Option<String> {
+	pub(super) fn get_body(&self) -> &Option<String> {
 		&self.body
 	}
 
-	pub(crate) fn get_file_stats(&self) -> &Option<Vec<FileStat>> {
+	pub(super) fn get_file_stats(&self) -> &Option<Vec<FileStat>> {
 		&self.file_stats
 	}
 }
